Implement Getting and Deleting users into service and repository
Adds DeleteUserByID, DeleteUserByUsername, GetUserByID, and GetUserByUsername to service and repository. Additionally adds DeleteUser to service. Tests for each are added too, with a MockUserRepository in service/user_test which captures parameter, allowing for tests to verify actions which modify the data store, such as create and delete (get is ignored as this is an implementation detail, probably?). Also, a 'factory' for random user generation in model/testutils. (This probably should've been more than 1 commit)